Moved migration helpers to fix migration error
