+++ Updates/4.12.4-b1.php
@@ -10,7 +10,7 @@
 namespace Piwik\Plugins\TagManager;
 
 use Piwik\Plugins\TagManager\Template\Variable\MatomoConfigurationVariable;
-use Piwik\Plugins\TagManager\Updates\NewVariableParameterMigrator;
+use Piwik\Plugins\TagManager\UpdateHelper\NewVariableParameterMigrator;
 use Piwik\Updater;
 use Piwik\Updates as PiwikUpdates;
